Rename `extension.ts` to `LanguageClientManager.ts`
